Componentize Person and Role html and functionality
